Update into to long

+++ shared/src/commonMain/kotlin/com/fitnessgym/ExercisesType.kt
@@ -1,10 +1,10 @@
 package com.fitnessgym
 
-sealed class ExercisesType(val literal: Int) {
+sealed class ExercisesType(val literal: Long) {
     data object SERIEA: ExercisesType(1)
     data object SERIEB: ExercisesType(2)
 
-    fun getName(literal: Int): String = when(literal) {
+    fun getName(literal: Long): String = when(literal) {
         SERIEA.literal -> "Serie A"
         SERIEB.literal -> "Serie B"
         else -> ""
